If the user is: 1) Using svn+ssh 2) Using ssh-agent (as detected by environment) 3) Hasn't set disable-agent-check,
kdesvn-build will now make sure that the ssh agent is actually maintaining a
list of identities, and if it isn't will run ssh-add for you.  Hopefully this
will go far towards fixing the problem where if you forgot to run ssh-add that
kdesvn-build would merrily sit and wait for your to type your passphrase.

svn path=/trunk/KDE/kdesdk/doc/scripts/kdesvn-build/; revision=431712