Correct some comments.
Add a default "kde-langauges" option (defaults to empty).  This is unimplemented but I will
try to implement this in the next week.

Correct the munge_lists() procedure to clear the checkout-only option correctly if more than
one module needed it.  (munge_lists() is used for the shortcut on the command line e.g.
doing kdesvn-build --svn-only kdelibs/khtml to update only kdelibs/khtml).

svn path=/trunk/KDE/kdesdk/scripts/kdesvn-build; revision=481684