Add preliminary CMake support to kdesvn-build, and update the sample file to suit.
This adds two new options: use-cmake (but this one will go away once CMake is the only
option for KDE 4), and cmake-options (but I think I will find a way to make it so you never
need to use this).

Also added is a new command line option, --run (which runs the given command using the
internal kdesvn-build environment).

The documentation has also been updated.

The check to see whether unsermake needs to be downloaded is also very slightly improved.

Finally, kdesvn-build does not set KDEDIR, it sets KDEDIRS.  Score one for removing deprecated
interfaces...

I think I've caught most of the places where the fact that we're using CMake is an issue
(i.e. do-not-compile, inst-apps, trying to make apidox, cmake-options adding to
global options like configure-flags does, etc.)  If I haven't, please e-mail me
or file a bug against kdesvn-build.

Seeing as how this is the first release with support for CMake, it should come as a surprise
that it may not work!  So, please test now so I can make a real 1.1 release in the next week
or so.

svn path=/trunk/KDE/kdesdk/doc/scripts/kdesvn-build/; revision=523429

Original commit: 73a544c8
https://invent.kde.org/sdk/kdesrc-build/-/commit/73a544c8b79655959666850f2e1be0ecc71eae4d