Try to be more intelligent with modules that have a different kdedir setting.
Now if a module has a different kdedir than kdelibs, the kdelibs kdedir setting will be
appended to the end of KDEDIRS.  i.e. if kdebase had kdedir = /usr and kdelibs had kdedir =
/home/kde/kde, the KDEDIRS for building kdebase would be "/usr:/home/kde/kde"

I'm hoping this helps to fix or at least identify the problem going on in bug 122843.

CCBUG:122843

svn path=/trunk/KDE/kdesdk/scripts/kdesvn-build; revision=526651