Add code to use the kdelibs4_snapshot for kdelibs if the 'branch kde4-snapshot' option is selected, instead of requiring weird hacks in the config file to use the snapshot.
e.g.

module kdelibs
  branch kde4-snapshot  # Use the branches/work/kdelibs4_snapshot stuff.
end module

Make the snapshot the default for kdelibs, which should help with the Coverity stuff.

Don't try to download a snapshot if override-url is in effect for a module.

Update the sample file to match, including a comment at the use-stable-kde option to warn
  the user to also update the kdelibs options.

Touch up the pretend mode output a bit.

svn path=/trunk/KDE/kdesdk/scripts/kdesvn-build; revision=529207