As requested by Dirk: Add support for automatically installing D-BUS 0.62.  Dirk, please review.  (It installs D-BUS to ${kdedir}).
Right now this is restricted to the case when kdesvn-build is being run as part of the
Coverity code scanning process.  For the rest of us, there is a perfectly fine D-BUS
release which you should install using your distribution's package managing tools, and
not with kdesvn-build.

CCMAIL:mueller@kde.org

svn path=/trunk/KDE/kdesdk/scripts/kdesvn-build; revision=556055