Fix build failure reported in the kdesvn-build 1.7 release.  Someday I'll have the time to do better testing, but this is still a fairly embarrassing bug to make it to a release. :(
Guess I should heed the motto of "Slow is Fast" and stop rushing myself.

The cause of the bug was moving around the definition of $script.  Later code in the function
(inappropriately) used the value of $script which used to coincidentally be correct.

As a bonus though the debugging for this also uncovered a long-hidden bug in the venerable
get_option subroutine which would cause qt-copy's configure to fail if the configure-flags were
empty.

Also persistent option reading is given a sane default, which should hopefully fix the other
build failures.

BUG:172288

svn path=/trunk/KDE/kdesdk/scripts/kdesvn-build; revision=868757