Make kdesrc-build skip building modules that have no source code update.
I've rejected this feature a couple of times due to the risk of having an inconsistent
build, but honestly this isn't a realistic probability for many of the possible incompatible
updates, and using --refresh-build is already sometimes required.

So since the patch provided by Rafael Fernández López was already in such good shape (and
provided out of the blue) I decided I'd finally relent and get the feature in.

So now kdesrc-build will detect whether actual updates occurred to the source and only try
to build in that situation. This can, of course, be overridden with --refresh-build.

CCMAIL:ereslibre@kde.org

svn path=/trunk/KDE/kdesdk/scripts/kdesrc-build; revision=1160411