Change how the nascent module-set grouping operates.
OK, so I had the idea last week to remove the repository name from the module-set
parameter, and add the ability to have the 'repository' option automatically use aliases
defined by git-repository-base. dfaure seemed to like that, and I've heard no other
negative feedback, so I've gone ahead and implemented that. (The old style is a syntax
error instead of me being nice and accepting it since it's never been part of a release).

So now, something like this should work:

module-set
  use-modules automoc akonadi attica # and etc.
  repository kde
end module-set

For a "kde" repository previously defined in a git-repository-base in the global section.
You can throw any other option into that module-set, and it will be applied to all of the
modules (that's been true for about a week now).

While I was at it, I fixed the source update for git modules to actually give
the "Updating" or "Cloning" module message similar to how svn modules are
noted, and I've added a warning if a module-set with no use-modules is declared.

I'd like for this to be all the new features before the next release, so please keep
testing!

svn path=/trunk/KDE/kdesdk/scripts/kdesrc-build; revision=1207654