Re-enable git-clone output for qt-copy.
I thought that using the -v flag to git would be sufficient to keep gitorious.org's
git daemon running right when cloning qt-copy, but that output also *must* be redirected
to the tty. This restores the kdesrc-build 1.12 behavior of outputting git modules clone
status to the tty, but only for qt-copy since otherwise gitorious.org's git daemon will
fail for some strange reason.

This hopefully should fix bug 262145, thanks to the reporter for being so proactive in
reporting the bug (including to gitorious) and testing changes and older versions of
kdesrc-build.

BUG:262145

svn path=/trunk/KDE/kdesdk/scripts/kdesrc-build; revision=1214740