Add CMake build support for kdesrc-build.
To avoid kdesrc-build build errors, it's probably easiest just to
have a shell CMakeLists.txt which CMake can process and do nothing
with.

At some point it will be used to make the docs build, but right
now the docs install to the wrong directory (as if the name of
the project was "doc" instead of kdesrc-build), so installation
is disabled.