Do not run git-diff in pretend mode.
I had left git-diff as system() instead of safe_system() (which
respects --pretend) when looking to see if git-stash needed to be run.

However, this can sometimes significantly slow down the kdesrc-build -p
run, and the only real benefit was not confusing users with git diff and
git diff --cached command lines.