Fix detection of whether git-stash is needed.
I broke the code that determines whether git-stash is determined in
commit 2b0d0d62481880ffc1e1f0e4820f2944b78610d3. The problem is that
the system() Perl function returns the exit code and miscellaneous
process exit data (such as whether the process failed to run, etc.)
and I actually needed that data.

In the commit I refer to, I switched to my safe_system() which only
returns the exit code and now kdesrc-build thought there was an error
running git, when instead git was returning the proper exit code for
a module with local changes. Thanks to dfaure for the report.

Not closing the bug here in case I merge this somewhere else later.

CCBUG:265325