Stop guessing tag names for kdesupport-for-4.foo
kdesupport used to have tags made up of the form kdesupport-for-4.2,
kdesupport-for-4.3, etc. which contained the releases of the various
underlying different libraries that were minimum requirements for the
corresponding KDE version.

Now that kdesupport is mostly moved to git, those tags are not being
created, so don't bother trying to guess them for 4.6 and greater.