Get closer to supporting XML modules from CLI.
This change refines module support a bit by introducing a Module object
for use by the rc-file and command line option-reading code. What's
needed next is to move the XML module expansion from parse_moduleset to
occur after process_arguments() and read_options() (but before the
actual rest of the process, of course).