Refactor XML searching.
kdesrc-build now supports building XML modules from the command line
(well, building is disabled in favor of debugging printouts, but the
module list gets setup right at least).

To force a command-line module to be thought of as an XML module,
precede the name with a +.

e.g. kdesrc-build --pretend +kde

is instructive at this point... ;)