Use correct value from git-repository-base.
I ported to use a key instead of a value.

In addition, properly tell whether a class method (for package Module)
is operating as an object method or as a class method.

I get a lot of errors about things not being a git module, which leads
me to believe that auto-detecting whether a module is svn or git is no
longer working, so test with care.