Maintain existing scm type for source directories.
With some git module IDs in the XML ending up the same as still-used
Subversion modules, it makes more sense to maintain the existing status
of present source directories instead of trying to use the wrong scm
tool.

Later I need to add detection for e.g. "MIGRATED" files to influence scm
type detection, but this works a bit better for now.