Prefer module-specific options to guess scm type.
kdesrc-build tries to figure out if a module is git-based or svn-based
automagically. Now the script will see if the 'repository' or
'svn-server' settings are set directly for a module and if so, those
factors will take precedence.

Otherwise kdesrc-build will still prefer an existing .svn directory to
guess svn, but in the future kdesrc-build will start tending to guess
git if there's no other info.