Replace use-xml-modules with repository overload.
Having use-modules and use-xml-modules is stupid.

Instead, you can use "kde-projects" for the existing repository option
to allow kdesrc-build to know that you mean for the module to be built
based on the KDE Projects structure, and kdesrc-build will handle the
rest, and then use use-modules as normal.

This means that you can't mix XML-based modules and the old-style
modules in the same module-set, but oh well...

You can still use +-prefixed module identifiers on the command line as
well.