Make unknown command-line modules an error.
It used to be that unknown modules would simply be attempted to build
anyways, and it would be amusing if it failed, but with the new method
of handling regular and XML-based modules, an unknown module simply
causes no output and a successful return message.

Now it results in an exception being thrown.