Allow source updates when using --resume-*.
As noticed by dfaure. It turns out that I was simply lazy and didn't
want to use the same logic in both get_update_list and get_build_list,
but those functions are effectively useless now anyways with the new way
of figuring out modules to work on (although there's still a few
bugfixes needed for that new code).