Save info on how the module list was generated.
i.e. was the module list given on the command line directly, or allowed
to be generated from the config file. It's only coarse-grained (e.g.
filtering options on the command line still count as modules coming from
the config file).

This will be used for properly handling the order for module uninstalls
and possibly for restoring old behavior on module installation with
--install.