Split out rc file finding and reading.
Now the act of actually finding the rc file to use is separate from the
act of reading it.

Again, this is mostly to actually make testing easier, although it will
make it very easy to add in a "real" built-in default config file
instead of the current programmatic hacks I have going on now.