Add a default Phonon backend to sample rc file.
Nowadays Phonon does not come with any backends built in, so we need to
specify what backend to use.

The two best choices are phonon-vlc and phonon-gstreamer. I've chosen
phonon-gstreamer for the default merely because that's the one that
works on my system.

Thanks to Ghislain MARY for bringing it to my attention.

BUG:263937