Fix one more reference to my home dir layout.
The logging subsystem test case was passing because of an old log
directory I still had on my system. I've fixed the test case to use the
existing temporary test directory and to not use any existing log
directories.

I've gone over the code one more time, this should hopefully be the last
reference to anything on my system. :-/