Add vim fold markers.
I try to keep kdesrc-build as a single integrated script for ease of
use, so it is extremely long by this point. I do need to go through it
and clean up, but I've read about a way to have vim automatically create
folds based on markers added to the source, which makes navigating the
file much easier.