Start on revamp of kdesrc-build code strcture.
So what I've done so far is to introduce a "BuildContext" class, since
it doesn't make sense for a simple Module class to also track the global
module list. This also makes the phase handling (also split out to a
different list to avoid code duplication) a little more normal.

Now the only global state in Module is related to the source of where
modules to be built was established (config or cmdline), which probably
needs to be moved as well.

In the long term much more will be modularized (even though I'll
probably continue to leave it all within one file).