Use libwww-perl instead of Net::HTTP/Net::FTP
Sorry about changing dependencies, but LWP seems to be the standard Perl
solution for downloading files, so it's likely to be more prevalent
anyways. Neither group of dependencies appears to be part of core Perl
so I'm having to pick either way.

Picking this way allows for a large reduction in code usage while still
properly handling redirects.