Actually use right method name.
Sometimes I really miss static typing. This is the kind of issue that is
supposed to be caught by unit tests in dynamic languages, but the test
suite isn't close to that extensive yet.

This fixes bug 268072 (kdesrc-build without modules passed in fails).
There is a test case which simulates kdesrc-build without modules passed
in, but it doesn't simulate the entire sequence and so didn't catch this
error. My apologies.

BUG:268072