Use proper branch when cloning git modules.
kdesrc-build only used the "branch" option for git modules when updating
existing git clones, not when actually cloning a module. Now it tries to
check out the appropriate branch immediately after cloning the module.

Thanks to kgw for reporting the bug. The fix should be in the eventual
1.14 release (or you can simply download the kdesrc-build script from
git)

BUG:268202
FIXED-IN:1.14