Add "include" option for config file reading.
Based on a recommendation from Michael Hansen, this commit adds the
ability to include other files from the rc file used for kdesrc-build.

Tilde-expansion is supported on the filename, but only of the ~/foo
variety (not ~user/foo).

It works by essentially intercepting the input stream as necessary, so
unfortunately things like a common "global" followed by different
"global" sections doesn't work (although let me know if that's needed,
it wouldn't be too hard).

On the other hand each rc-file would be able to include the global
options that are common.

To use, simply use "include /path/to/file-to-include" as a single line
in your config file. Leading spaces are supported, but the remainder of
the file after the "include" and all whitespace immediately following
"include" is assumed to be the filename, so don't leave trailing spaces.

BUG:223331
FIXED-IN:1.14