Make kdesrc-build control the 'origin' remote.
I've received a few comments by now that developers expect kdesrc-build
to automatically reposition the 'origin' remote as appropriate for
whatever the repository setting in the rc file is. I had avoided that
based on not touching things that the developer controls, but it seems
that developers expect the opposite in general, so I'm happy to do that
too.