kdesrc-build: Remove global update_list.
In keeping with my goal of better modularization, remove the global
@update_list, in favor of the global context object which should be
passed to methods that need that information.

Many places still just generate the appropriate update list manually
that probably don't need to, but that's fine for now.

The long-term goal is for module *objects* to be used in most places
instead of just passing the module name around, but that will come
gradually.

This does mean removing the Coverity plugin support (I left in the code
that tries to do less if Coverity is running), but I do not believe
Coverity is using kdesrc-build to build KDE at this point. If this is
not the case let me know and I'll see what I can do to maintain or
update compatibility for that.