Only try to download kde_projects.xml once per run.
After my internet went down last night I noticed that the error message
I had the next day didn't show up right (apparently the HTTP::Response
doesn't stringify like I thought.

Trying to confirm that was fixed led me to realize the way I handle
downloading the XML is overly complicated in the first place. Now we
only ever try to download the file once, and if that fails we delete the
result so we can't use it (much like make on failed object file builds).

The other thing we do in case an XML file that is corrupt is present on
disk is to always try downloading the XML file over the existing one on
non-pretend builds, which should hopefully ensure it's fixed over time.