Remove post-build error emailing.
kdesrc-build has supported emailing error messages for some time, at the
conclusion of the build process. There is no easier way to "modularize"
this code than to elide it completely, and given that no one seemed to
use it in the brief poll I conducted, I've removed it.

Original commit: e7e89304
https://invent.kde.org/sdk/kdesrc-build/-/commit/e7e89304e9e583d017a5ebaed3a836cae7b61170