Add accessor for "global" module data.
Normally kdesrc-build uses the 'global' pseudo-module in order to store
config data which is generic across different build modules. To be able
to move "get_option" to Module we need a way to handle that, and to
allow methods to operate on either generic or specific options we need
to ensure that whatever method is used has a common interface to Module.

The method I chose is to have a kind of parallel Module created which is
held by the BuildContext, but not in the normal list of Modules. This
commit creates that globalModule and adds an accessor for it.