Add way to lookup Module objs by name.
It is sometimes needed to be able to determine information about a
module based just on its module name. It used to be the module name was
all the information that a function had, and so it was sufficient by
definition, but with Module objects that is not always the case.

To allow for finding the matching Module for a module name add a search
function to the BuildContext.