Ensure only one BuildContext is in use.
Using dclone() on passed in modules in updateModulePhases causes the
Module's buildContext to also be cloned, but the function calling
updateModulePhases is working on an entirely different BuildContext, and
this inconsistency can cause problems in the newly-introduced
lookupModule function.

Since there's no reason to protect the input modules, just don't bother
using dclone.