Port svn_module_url and deps to use Module.
Modularization continues. A couple of functions actually really only
need to deal with the string value of the name. These functions have
been changed to accept the name into a param named "moduleName" since
I'm trying to ensure once this is done that '$module' will always
actually be a Module.

Note this uses the 2-argument getOption introduced in the last commit.