Allow for specifying all subprojects of a project.
Modules like kdegraphics are both the ancestor of a group of logical
child modules, and repos in their own right. kdegraphics is marked as
non-active since it's repo just exists to allow for building the
subprojects, but that means kdesrc-build won't build it.

So, allow for specifying modules *under* a given module. This allows for
a syntax as follows to work:

module-set kdegraphics-set
    repository kde-projects
    use-modules kdegraphics/libs kdegraphics/*
end module-set

(The kdegraphics/libs forces those modules to be added first. They will
not be readded by the later kdegraphics/*)