Move man page to DocBook, install it.
asciidoc was a nice start, but modifying the generated XML is not too
bad (and must be done to comply with KDE documentation guidelines
anyways).

So, remove the asciidoc and instead generate man pages from DocBook XML.

I've taken the liberty of using the more specific content tags but the
overall content should be approximately the same.