Use more "standard" Perl for kde: auto-handling.
Laurent Montel reported an error that indicates a syntax error in the
git error handler for the kde: support. Even after extracting out the
error handler and forcing it to run Perl works just fine here, so it's
not a semantic issue either.

So instead, just use a plain 'ol Perl 5.0-style hash to figure out the
error message, no need to make it complicated.

In addition I added the stderr > /dev/null redirection that I forgot to
have in place before.