kdesrc-build: Prefer module type to on-disk info.
This commit makes the "module type" as determined from the
{repository,svn-server} options be the preferred method of determining
which scm tool to use.

This is required to allow nuking the source directory for old checkouts
in an upcoming commit (which was itself made an issue with the 1.14
release, that reorganized the on-disk layout, causing older kde-baseapps
checkouts with konsole/kate subdirs to conflict with the git
konsole/kate).