Add --delete-my-patches option.
This doesn't so much actively search for things to delete. Instead it's
a required permission from the user to remove source directories (which
is similar to the remove-after-install option).

Right now this will only be used when trying to clone a git module and
the destination source directory already exists somehow, which has
happened for a lot of 1.14 early adopters due to old kde-baseapps
checkouts having a kate or konsole directory that conflict with where
kdesrc-build wants to put the current kate/konsole directories.