Correct the XML::Parser char handler.
Ralf Jung noted to me in an email that the SAX-style parser used for the
KDE projects database does not handle the condition that the char
handler is called more than once for a tag, giving the akonadi
repository setting as an example of a problem for him.

And what do you know, a bug that's reproduceable here! I'm not sure why
the parser splits up the akonadi repo URL for both of us but this is my
fault so I suppose it doesn't matter. Ralf submitted an example fix,
which I simplified down to this. Thanks for the detailed report and
corrections to fix the issue!

CCMAIL:ralfjung-e@gmx.de