Large refactoring of build process.
Should be mostly the same as far as operation goes, but in this commit I
introduce a build system mixin similar to the scm handler mixin. There's
still more code that should be moved but we're getting closer to having
a simple build process where all the scm-specific stuff doesn't clutter
the main code with the equivalent of #ifdefs everywhere.