Improve pretend handling.
2 changes related to handling --pretend + no-source-dir.

1. svnInfo:
If we can't change to the source dir before running svn info that should
be an error, so treat it as such.

Except, in pretend mode that's OK... so just check for that and return
some nonsense to keep things going.

2. Qt configure:
Again if there's no source dir due to pretending, just move on.