Simplify safe_make a great deal.
Prior to this, safe_make was handling a lot of different things that
unnecessarily cluttered it up (e.g. determining whether it was safe to
use make-options or not based solely on the args parameter, determining
whether to use make-install-prefix, etc.)

Now the actual callers of safe_make (where it is already known whether
or not to use make-options or make-install-prefix) pass that kind of
information in.

Even with an increase in code at the calling sites and the more
extensive comments, the simplifications in safe_make allow for an
overall decrease in code size.