Improve split_quoted_on_whitespace.
Apparently Perl has included (since version 5.0, no less) a module that
handles almost the exact case split_quoted_on_whitespace exists for.

The impetus for this is that the '${MODULE}' syntax for dest-dir is
apparently broken due to problems with the mixture of braces and quotes
in split_quoted_on_whitespace.

This change does not break the test suite (but does add a relevant test)
and does seem to fix the issue with strings with braces.

CCMAIL:rakuco@FreeBSD.org