Define progress output option for build systems.
This actually is already present, but is checked for by looking for
cmake buildsystems directly, whereas this determination should really
belong to the build system (now that such an entity is defined).

Note that it's not supported with l10n from what I understand, even
though it uses CMake... though this is just me porting the old behavior,
it may work fine now.