Support kde-build-metadata ignore info.
This unfortunately isn't really the best way to do it (in retrospect
this should be handled in ensure_projects_xml_present since we want the
metadata whenever we'd be downloading the XML anyways). Plus it really
probably isn't necessary to subclass KDEProjectUpdate.

However, the perfect is the enemy of the good and this seems to be
working for me now and I will be away on holidays for a week or so, so
best to get it in now, imperfect as it is.

Let me know if this doesn't work, also keep in mind the
'kde-build-metadata' repo is open for business if there's other modules
that should be ignored! ;)

BUG:288378