Modify getOption check for Qt to be generic.
Now as long as the build system is 'QtBuildSystem' we will ignore
Qt-specific options.

One problem is that the order of evaluation *had* to be reversed by
doing this, the way I have it coded now ends up with some kind of
5-level mutual recursion since the build system check will try to lookup
what directory to use if it's not already configured, and that
necessarily involves getOption calls.

Reversing the order of evaluation allows us to defer this check until
after it's actually necessary (by which point no special lookup should
be needed to figure out the build system type) but there's probably a
better way to do this...