Do not require the Qt module to be named 'qt-copy'.
Pretty much what it says in the subject, although it depends at this
point on auto-detecting the proper build system correctly.

If you want to see what build system kdesrc-build thinks is required,
run with --verbose or --debug, and the build system will be displayed
with the module name during the build phase.

I mistakenly said an earlier commit implemented this feature though,
it's actually *this* commit.

CCBUG:265255