Move install to Module and build system.
Right now that just means the functionality is split between
Module::install and GenericBuildSystem::installInternal but it's now at
least possible in theory for e.g. the qmake build system to add the
required DESTDIR= parameter to install to the right location.

This has built for me just fine on a couple of modules so far, haven't
been able to compile further though.