Get l10n install closer to working.
This gets it far enough into the install subroutine to get
installInternal called. The issue is just that the language module
itself is a grouping of languages to build, each of which is really its
own module.

So the default file the install procedure looks for to tell if the build
system was setup won't work, as it's one level too high. (e.g. it looks
for $builddir/Makefile instead of $builddir/$lang/Makefile).

With this change it's possible we'd try to install a language that
didn't actually build at this point, but that's better than not trying
at all.