Restore l10n build functionality for default build.
My previous patches required you to pass 'l10n' on the command line, now
kdesrc-build will add the required module to the build automatically if
the kde-languages option is in use (at the end of the build).

This should complete making the kde-languages option work, but more
testing is always good! ;)

BUG:283027