Auto-dereference list refs only if really refs.
This allows single-elements lists to not be guessed as a list reference
in list_has, which helps with bug 291357... but still doesn't explain
why old code that used list_has didn't have the same issue.

I have included a test case which fails and is likely related to the
reason though, so hopefully won't be hard to pin down from here.

CCBUG:291357