Don't double-add function args to environment.
Although dumber than I thought the bug would be, it at least was easy to
find.

The reason list_has() accepted the prior code was that it seemed to be
getting called with a @curList of 0 entries instead of 1 entry (when
there was no existing environment variable set).

I'm still not entirely sure but I believe that was caused by the bug
fixed by this commit, where the environment variable name would end up
prepended to the value. If there was no value assigned it would just
assign a value with 1 entry, which would have tripped the bug when used
later.

Please let me know if this is fixed, I've tried to add a few more
testcases which all pass at this point.

BUG:291357