Simplify list_has yet again.
Must of my problem tracking down the last commit was with what I thought
was another bug in the list_has util method.

Investigating that has led me to the conclusion that prototype usage
here is just a bad idea, so I'm picking a calling convention that's
mostly like to cause a runtime error if violated (i.e. passing the list
as a reference) and will be using it consistently.