"Mirror" project database instead of always downloading.
LWP supports a "mirror" function that forgoes downloading a file it's
already present and unchanged on the server (checked using
If-Modified-Since).

The features kdesrc-build uses LWP to download files for are all either
improved by using mirror or indifferent to it, so switch to avoid having
KDE infrastructure always have to send the same bits over the wire.